+++ packages/web-components/fast-foundation/src/number-field/number-field.ts
@@ -119,13 +119,22 @@ export class NumberField extends FormAssociatedNumberField {
      */
     @attr({ converter: nullableNumberConverter })
     public max: number;
-    public maxChanged(previousValue, nextValue): void {
-        this.max = Math.max(nextValue, this.min ?? nextValue);
+
+    /**
+     * Ensures that the max is greater than the min and that the value
+     *  is less than the max
+     * @param previous - the previous max value
+     * @param next - updated max value
+     *
+     * @internal
+     */
+    public maxChanged(previous: number, next: number): void {
+        this.max = Math.max(next, this.min ?? next);
         const min = Math.min(this.min, this.max);
         if (this.min !== undefined && this.min !== min) {
             this.min = min;
         }
-        this.valueChanged(this.value, this.value);
+        this.value = this.getValidValue(this.value);
     }
 
     /**
@@ -136,16 +145,26 @@ export class NumberField extends FormAssociatedNumberField {
      */
     @attr({ converter: nullableNumberConverter })
     public min: number;
-    public minChanged(previousValue, nextValue): void {
-        this.min = Math.min(nextValue, this.max ?? nextValue);
+
+    /**
+     * Ensures that the min is less than the max and that the value
+     *  is greater than the min
+     * @param previous - previous min value
+     * @param next - updated min value
+     *
+     * @internal
+     */
+    public minChanged(previous: number, next: number): void {
+        this.min = Math.min(next, this.max ?? next);
         const max = Math.max(this.min, this.max);
         if (this.max !== undefined && this.max !== max) {
             this.max = max;
         }
-        this.valueChanged(this.value, this.value);
+        this.value = this.getValidValue(this.value);
     }
 
     /**
+     * The default slotted items
      * @internal
      */
     @observable
@@ -158,29 +177,56 @@ export class NumberField extends FormAssociatedNumberField {
     public control: HTMLInputElement;
 
     /**
-     *
-     * @param previousValue - previous stored value
-     * @param nextValue - value being updated
+     * Flag to indicate that the value change is from the user input
+     * @internal
      */
-    public valueChanged(previousValue, nextValue): void {
-        let value: number | string = parseFloat(nextValue);
-        if (isNaN(value)) {
-            value = "";
-        } else {
-            value = Math.min(value, this.max ?? value);
-            value = Math.max(value, this.min ?? value);
+    private isUserInput: boolean = false;
+
+    /**
+     * Validates that the value is a number between the min and max
+     * @param previous - previous stored value
+     * @param next - value being updated
+     * @param updateControl - should the text field be updated with value, defaults to true
+     * @internal
+     */
+    public valueChanged(previous: string, next: string): void {
+        this.value = this.getValidValue(next);
+
+        if (next !== this.value) {
+            return;
         }
 
-        this.value = value.toString();
-
-        if (this.proxy instanceof HTMLInputElement) {
-            this.proxy.value = this.value;
+        if (this.control && !this.isUserInput) {
+            this.control.value = this.value;
         }
 
-        if (previousValue !== undefined) {
+        super.valueChanged(previous, this.value);
+
+        if (previous !== undefined && !this.isUserInput) {
             this.$emit("input");
             this.$emit("change");
         }
+
+        this.isUserInput = false;
+    }
+
+    /**
+     * Sets the internal value to a valid number between the min and max properties
+     * @param value - user input
+     * @param updateControl - should the text field update to the valid value
+     *
+     * @internal
+     */
+    private getValidValue(value: string): string {
+        let validValue: number | string = parseFloat(parseFloat(value).toPrecision(12));
+        if (isNaN(validValue)) {
+            validValue = "";
+        } else {
+            validValue = Math.min(validValue, this.max ?? validValue);
+            validValue = Math.max(validValue, this.min ?? validValue).toString();
+        }
+
+        return validValue;
     }
 
     /**
@@ -201,7 +247,6 @@ export class NumberField extends FormAssociatedNumberField {
             : 0;
 
         this.value = stepUpValue.toString();
-        this.control.value = this.value;
     }
 
     /**
@@ -222,10 +267,10 @@ export class NumberField extends FormAssociatedNumberField {
             : 0;
 
         this.value = stepDownValue.toString();
-        this.control.value = this.value;
     }
 
     /**
+     * Sets up the initial state of the number field
      * @internal
      */
     public connectedCallback(): void {
@@ -248,6 +293,7 @@ export class NumberField extends FormAssociatedNumberField {
      */
     public handleTextInput(): void {
         this.control.value = this.control.value.replace(/[^0-9\-+e.]/g, "");
+        this.isUserInput = true;
         this.value = this.control.value;
     }