+++ packages/tooling/fast-figma-plugin-msft/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @microsoft/fast-figma-plugin-msft
 
-This log was last generated on Sun, 07 Nov 2021 07:12:36 GMT and should not be manually modified.
+This log was last generated on Wed, 10 Nov 2021 07:12:54 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.8.16
+
+Wed, 10 Nov 2021 07:12:54 GMT
+
+### Patches
+
+- Bump @microsoft/fast-components to v2.16.1 (robarb@microsoft.com)
+
+## 0.8.16
+
+Tue, 09 Nov 2021 07:14:09 GMT
+
+### Patches
+
+- Bump @microsoft/fast-components to v2.16.0 (jes@microsoft.com)
 
 ## 0.8.16